reuse the cli arg parsing from rust for python
